refactor: name the files after what a developer types
The layout had grown by accretion and no longer matched the CLI. main.go was
916 lines of dispatch plus help; the whole iOS signing-team feature lived in
doctor_fix.go, which is about repairing an Xcode install; commands.go held
three unrelated ones. Finding the code behind a command meant grepping.

Files are now named for the surface:

  main.go                 dispatch only
  help.go                 the help text for every command
  cmd_<command>.go        one per command — new, build, dev, app, tools, ...
  <platform>_build.go     how a platform is built and run
  <platform>_package.go   how it is packaged for its store
  <thing>_toolchain.go    provisioning: android, tailwind
  host_packages.go        brew/apt/pacman dependencies

Platform names lead rather than trail, and that is not style: build_ios.go and
build_android.go compile only for GOOS=ios and GOOS=android under Go's implicit
build constraints. That silently excluded two files from every build earlier in
this work, so the rule is written down in main.go where the next person will
see it.

Pure moves — no behaviour changed, and the command surface was re-run to
confirm it.
