+++ cmd/irgo/cmd_clean.go
