+++ cmd/irgo/cmd_tools_remove.go
